Renamed from SearchEngine to SearchIndexer as a more appropriate name, added jar build to main pom, introduced commons cli and basic commands
