+++ BusinessLogic/pom.xml
@@ -2,8 +2,8 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
     <parent>
-        <groupId>dk.easv.SearchEngine</groupId>
-        <artifactId>SearchEngine-Main</artifactId>
+        <groupId>dk.easv.SearchIndexer</groupId>
+        <artifactId>Search-Indexer</artifactId>
         <version>1.0</version>
     </parent>
     <artifactId>BusinessLogic</artifactId>
@@ -19,13 +19,13 @@
     
     <dependencies>
         <dependency>
-            <groupId>dk.easv.SearchEngine</groupId>
+            <groupId>dk.easv.SearchIndexer</groupId>
             <artifactId>DataAccessLogic</artifactId>
             <version>${project.version}</version>
         </dependency>
         
         <dependency>
-            <groupId>dk.easv.SearchEngine</groupId>
+            <groupId>dk.easv.SearchIndexer</groupId>
             <artifactId>DomainModels</artifactId>
             <version>${project.version}</version>
         </dependency>
