--- BusinessLogic/src/main/java/dk/easv/SearchEngine/BusinessLogicLayer/TermSplitterImpl.java
@@ -1,25 +1,0 @@
-package dk.easv.SearchEngine.BusinessLogicLayer;
-
-import dk.easv.SearchEngine.DomainModels.Term;
-
-import java.util.List;
-import java.util.stream.Collectors;
-import java.util.stream.Stream;
-
-public class TermSplitterImpl implements TermSplitter {
-
-    private final String regexToSplitOn;
-
-    public TermSplitterImpl(String regexToSplitOn) {
-        this.regexToSplitOn = regexToSplitOn;
-    }
-
-    @Override
-    public List<Term> splitLines(List<String> lines) {
-        return lines.stream()
-                .flatMap(line -> Stream.of(line.split(regexToSplitOn)))
-                .map(String::toLowerCase)
-                .map(Term::new)
-                .collect(Collectors.toList());
-    }
-}
