+++ DomainModels/module_domainmodels.xml
@@ -1,0 +1,94 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project name="module_domainmodels" default="compile.module.domainmodels">
+  <dirname property="module.domainmodels.basedir" file="${ant.file.module_domainmodels}"/>
+  
+  <property name="module.jdk.home.domainmodels" value="${project.jdk.home}"/>
+  <property name="module.jdk.bin.domainmodels" value="${project.jdk.bin}"/>
+  <property name="module.jdk.classpath.domainmodels" value="${project.jdk.classpath}"/>
+  
+  <property name="compiler.args.domainmodels" value="-encoding UTF-8 -source 8 -target 1.8 ${compiler.args}"/>
+  
+  <property name="domainmodels.output.dir" value="${module.domainmodels.basedir}/target/classes"/>
+  <property name="domainmodels.testoutput.dir" value="${module.domainmodels.basedir}/target/test-classes"/>
+  
+  <path id="domainmodels.module.bootclasspath">
+    <!-- Paths to be included in compilation bootclasspath -->
+  </path>
+  
+  <path id="domainmodels.module.production.classpath">
+    <path refid="${module.jdk.classpath.domainmodels}"/>
+    <path refid="library.maven:_com.google.guava:guava:19.0.classpath"/>
+    <path refid="library.maven:_org.mockito:mockito-core:1.10.19.classpath"/>
+  </path>
+  
+  <path id="domainmodels.runtime.production.module.classpath">
+    <pathelement location="${domainmodels.output.dir}"/>
+    <path refid="library.maven:_com.google.guava:guava:19.0.classpath"/>
+    <path refid="library.maven:_org.hamcrest:hamcrest-core:1.3.classpath"/>
+    <path refid="library.maven:_org.mockito:mockito-core:1.10.19.classpath"/>
+    <path refid="library.maven:_org.objenesis:objenesis:2.1.classpath"/>
+  </path>
+  
+  <path id="domainmodels.module.classpath">
+    <path refid="${module.jdk.classpath.domainmodels}"/>
+    <pathelement location="${domainmodels.output.dir}"/>
+    <path refid="library.maven:_com.google.guava:guava:19.0.classpath"/>
+    <path refid="library.maven:_junit:junit:4.12.classpath"/>
+    <path refid="library.maven:_org.hamcrest:hamcrest-core:1.3.classpath"/>
+    <path refid="library.maven:_org.mockito:mockito-core:1.10.19.classpath"/>
+    <path refid="library.maven:_org.objenesis:objenesis:2.1.classpath"/>
+  </path>
+  
+  <path id="domainmodels.runtime.module.classpath">
+    <pathelement location="${domainmodels.testoutput.dir}"/>
+    <pathelement location="${domainmodels.output.dir}"/>
+    <path refid="library.maven:_com.google.guava:guava:19.0.classpath"/>
+    <path refid="library.maven:_junit:junit:4.12.classpath"/>
+    <path refid="library.maven:_org.hamcrest:hamcrest-core:1.3.classpath"/>
+    <path refid="library.maven:_org.mockito:mockito-core:1.10.19.classpath"/>
+    <path refid="library.maven:_org.objenesis:objenesis:2.1.classpath"/>
+  </path>
+  
+  
+  <patternset id="excluded.from.module.domainmodels">
+    <patternset refid="ignored.files"/>
+  </patternset>
+  
+  <patternset id="excluded.from.compilation.domainmodels">
+    <patternset refid="excluded.from.module.domainmodels"/>
+  </patternset>
+  
+  <path id="domainmodels.module.sourcepath">
+    <dirset dir="${module.domainmodels.basedir}">
+      <include name="src/main/java"/>
+    </dirset>
+  </path>
+  
+  
+  <target name="compile.module.domainmodels" depends="compile.module.domainmodels.production,compile.module.domainmodels.tests" description="Compile module DomainModels"/>
+  
+  <target name="compile.module.domainmodels.production" depends="register.custom.compilers" description="Compile module DomainModels; production classes">
+    <mkdir dir="${domainmodels.output.dir}"/>
+    <javac2 destdir="${domainmodels.output.dir}" debug="${compiler.debug}" nowarn="${compiler.generate.no.warnings}" memorymaximumsize="${compiler.max.memory}" fork="true" executable="${module.jdk.bin.domainmodels}/javac">
+      <compilerarg line="${compiler.args.domainmodels}"/>
+      <bootclasspath refid="domainmodels.module.bootclasspath"/>
+      <classpath refid="domainmodels.module.production.classpath"/>
+      <src refid="domainmodels.module.sourcepath"/>
+      <patternset refid="excluded.from.compilation.domainmodels"/>
+    </javac2>
+    
+    <copy todir="${domainmodels.output.dir}">
+      <fileset dir="${module.domainmodels.basedir}/src/main/java">
+        <patternset refid="compiler.resources"/>
+        <type type="file"/>
+      </fileset>
+    </copy>
+  </target>
+  
+  <target name="compile.module.domainmodels.tests" depends="register.custom.compilers,compile.module.domainmodels.production" description="compile module DomainModels; test classes" unless="skip.tests"/>
+  
+  <target name="clean.module.domainmodels" description="cleanup module">
+    <delete dir="${domainmodels.output.dir}"/>
+    <delete dir="${domainmodels.testoutput.dir}"/>
+  </target>
+</project>
