+++ pom.xml
@@ -1,12 +1,12 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
-    <groupId>dk.easv.SearchEngine</groupId>
-    <artifactId>SearchEngine-Main</artifactId>
+    <groupId>dk.easv.SearchIndexer</groupId>
+    <artifactId>Search-Indexer</artifactId>
     <version>1.0</version>
     <packaging>pom</packaging>
     
-    <name>Search Engine</name>
+    <name>Search Indexer</name>
     
     <modules>
         <module>DataAccessLogic</module>
@@ -33,5 +33,36 @@
             <version>1.10.19</version>
         </dependency>
     </dependencies>
+
+    <build>
+        <finalName>search-indexer</finalName>
+        <plugins>
+            <!-- Set a JDK compiler level -->
+            <plugin>
+                <groupId>org.apache.maven.plugins</groupId>
+                <artifactId>maven-compiler-plugin</artifactId>
+                <version>2.3.2</version>
+                <configuration>
+                    <source>1.8</source>
+                    <target>1.8</target>
+                </configuration>
+            </plugin>
+
+            <!-- Make this jar executable -->
+            <plugin>
+                <groupId>org.apache.maven.plugins</groupId>
+                <artifactId>maven-jar-plugin</artifactId>
+                <configuration>
+                    <archive>
+                        <manifest>
+                            <!-- Jar file entry point -->
+                            <mainClass>dk.easv.SearchIndexer.Interface.MainApplication</mainClass>
+                        </manifest>
+                    </archive>
+                </configuration>
+            </plugin>
+
+        </plugins>
+    </build>
 
 </project>
